moved pin control functions to separate file
